working sessions and google login/logout
