+++ client/src/components/buttons/LoginButton.js
@@ -20,11 +20,7 @@ function LoginButton(props) {
   }
 
   function logout() {
-    window.location = 'https://localhost:5000/google/logout';
-    // fetch('https://localhost:5000/google/logout')
-    //   .then(response => response.json())
-    //   .then(result => props.updateSession(result))
-    //   .catch(err => console.log(err));
+    window.location = 'https://localhost:5000/logout';
   }
 
   return (
