--- server/requirements.txt
@@ -1,7 +1,0 @@
-requests==2.25.1
-Flask==1.1.2
-Flask-Cors==1.10.3
-Flask-Login==0.5.0
-Flask-WTF==0.14.3
-oauthlib==3.1.0
-pyOpenSSL==20.0.1
